# Calculate intermediary derivatives during forward pass and optimize/simplify graph layout