--- a/README.md
+++ b/README.md
@@ -1,11 +1,5 @@
 # ReverseDiffPrototype
 
-This is a prototype Julia implementation of reverse mode AD, mainly for my own educational purposes.
+This is a prototype Julia implementation of reverse mode AD.
 
-I'm using https://github.com/pearsonlab/rdiff_proto as a reference implementation.
-
-<!-- [![Build Status](https://travis-ci.org/jrevels/ReverseDiffPrototype.jl.svg?branch=master)](https://travis-ci.org/jrevels/ReverseDiffPrototype.jl)
-
-[![Coverage Status](https://coveralls.io/repos/jrevels/ReverseDiffPrototype.jl/badge.svg?branch=master&service=github)](https://coveralls.io/github/jrevels/ReverseDiffPrototype.jl?branch=master)
-
-[![codecov.io](http://codecov.io/github/jrevels/ReverseDiffPrototype.jl/coverage.svg?branch=master)](http://codecov.io/github/jrevels/ReverseDiffPrototype.jl?branch=master) -->
+[![Build Status](https://travis-ci.org/jrevels/ReverseDiffPrototype.jl.svg?branch=master)](https://travis-ci.org/jrevels/ReverseDiffPrototype.jl)