--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -2,34 +2,51 @@ using ReverseDiffPrototype
 using Base.Test
 using ForwardDiff
 
+const RDP = ReverseDiffPrototype
+
+##################################################
+
 x = rand(5)
 out = zeros(x)
-testf1(x) = (exp(x[1]) + log(x[3]) * x[4]) / x[5]
-g1! = ReverseDiffPrototype.gradient(testf1)
 
-@test_approx_eq g1!(out, x) ForwardDiff.gradient(testf1, x)
+test1(x) = (exp(x[1]) + log(x[3]) * x[4]) / x[5]
+
+@test_approx_eq RDP.gradient!(out, test1, x) ForwardDiff.gradient(test1, x)
+
+##################################################
 
 x = rand(2)
 out = zeros(x)
-testf2(x) = x[1]*x[2] + sin(x[1])
-g2! = ReverseDiffPrototype.gradient(testf2)
 
-@test_approx_eq g2!(out, x) ForwardDiff.gradient(testf2, x)
+test2(x) = x[1]*x[2] + sin(x[1])
 
-const N = 10
-x = rand(2N^2 + N)
-out = zeros(x)
-function testf3(x)
-    k = length(x)
-    A = reshape(x[1:N^2], N, N)
-    B = reshape(x[N^2 + 1:2N^2], N, N)
-    c = x[2N^2+1:end]
-    return trace(log(A * B .+ c))
+@test_approx_eq RDP.gradient!(out, test2, x) ForwardDiff.gradient(test2, x)
+
+##################################################
+
+n = 2
+x = collect(1:(2n^2 + n))
+t = RDP.trace_input_array(x, Int)
+out = zeros(Float64, length(x))
+
+function generate_test3(n)
+    return x -> begin
+        @assert length(x) == 2n^2 + n
+        A = reshape(x[1:n^2], n, n)
+        B = reshape(x[n^2 + 1:2n^2], n, n)
+        C = x[2n^2+1:end]
+        return trace(log(A * B .+ C))
+    end
 end
 
-g3! = ReverseDiffPrototype.gradient(testf3)
+test3 = generate_test3(n)
 
-@test_approx_eq g3!(out, x) ForwardDiff.gradient(testf3, x)
+@test_approx_eq RDP.gradient!(out, test3, x) ForwardDiff.gradient(test3, x)
+
+##################################################
+
+x = rand(100)
+out = similar(x)
 
 function rosenbrock(x)
     a = one(eltype(x))
@@ -40,22 +57,19 @@ function rosenbrock(x)
     end
     return result
 end
-g4! = ReverseDiffPrototype.gradient(rosenbrock)
-x = rand(100)
-out = similar(x)
 
-@test_approx_eq g4!(out, x) ForwardDiff.gradient(rosenbrock, x)
+@test_approx_eq RDP.gradient!(out, rosenbrock, x) ForwardDiff.gradient(rosenbrock, x)
 
-# map of univariates
-x = rand(49) # using `rand(1:10, 49)` will cause this test to fail
-out = zeros(x)
-aux_fn5(x) = sqrt(abs(x) + x^2) * 0.5
-function testf5(x)
+##################################################
+
+x = rand(1:10, 49)
+out = zeros(Float64, 49)
+
+function test5(x)
     k = length(x)
     N = Int(sqrt(k))
     A = reshape(x, N, N)
-    return sum(map(aux_fn5, A))
+    return sum(map(n -> sqrt(abs(n) + n^2) * 0.5, A))
 end
 
-g5! = ReverseDiffPrototype.gradient(testf5)
-@test_approx_eq g5!(out, x) ForwardDiff.gradient(testf5, x)
+@test_approx_eq RDP.gradient!(out, test5, x) ForwardDiff.gradient(test5, x)