This PR changes our approach with the goal of making the backwards pass (nearly) non-allocating. We no longer need external graph storage or caching; the number type itself is now its own node in the graph, and maintains its own references to its parents. Since there's no longer a tape, I've named the new value type `TraceReal`. By using `RefValue`s for the `adjoint` field, I was able to make `TraceReal`s immutable.

The backwards pass for gradients now does a BFS starting at the result node. I use the normal queue-based implementation of a BFS, which means there is a small amount of allocation, but it's pretty negligible compared to allocation of the forward pass:

```julia
julia> include("test/rosenbrock.jl")
  0.058133 seconds (3.20 M allocations: 96.138 MB) # forward pass only
  0.154665 seconds (10 allocations: 896 bytes) # backward pass only
  0.214749 seconds (3.20 M allocations: 96.130 MB) # call to gradient! (forward + backward pass)
done
```

The majority of the time (~ 70%) of the backwards pass is graph traversal, rather than derivative propagation. Here's the same run, with the derivative propagation commented out:

```julia
julia> include("test/rosenbrock.jl")
  0.058232 seconds (3.20 M allocations: 96.130 MB)  # forward pass only
  0.128279 seconds (10 allocations: 896 bytes) # backward pass only, no derivative propagation
  0.198831 seconds (3.20 M allocations: 96.130 MB) # call to gradient! (forward + backward pass)
done
```

 To speed up the BFS, I played around with `DataStructures.Queue`, as well as rolling my own fixed-size queue using a custom ring buffer type, but neither were as performant as just `pop!`/`unshift!` on `Vector`. The queue never exceeds a size of ~20 for rosenbrock, so the `Vector` probably reaches full size via dynamic allocation rather quickly in most cases.

Inspired by @mlubin's approach in ReverseDiffSparse, intermediary derivatives are now calculated in the forward pass using the `ForwardDiff.Dual` type, which means the backwards pass is merely multiplication/addition. I haven't yet explored how far I could go with ReverseDiffSparse's traversal optimizations, which rely on the target function being data independent (i.e. no value-based branching) and representable as a tree (correct me if I'm wrong, Miles). It would be cool to optionally allow the user to annotate things or provide keyword arguments that would make these optimizations possible; I'll open an issue for this.

@jmxpearson I've also added you to the license.

I still have to figure out the right way to do array operation optimizations (e.g. `map` overloading) using this new graph structure.