@jrevels 2016-07-11
> I still have to figure out the right way to do array operation optimizations (e.g. map overloading) using this new graph structure.

My naive first thought on this is that we'll have to implement a `TraceArray` wrapper for `AbstractArray`s.