@jmxpearson 2016-07-12
Thanks. I finally got the big PR we'd been working on settled, so I'll have
more time to take a look at this.

On Mon, Jul 11, 2016 at 1:30 PM Jarrett Revels <notifications@github.com>
wrote:

> Build is failing because I need to tag ForwardDiff master.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/jrevels/ReverseDiffPrototype.jl/pull/10#issuecomment-231805225>,
> or mute the thread
> <https://github.com/notifications/unsubscribe/ADBM57rRWM4GKUHR1H-FzrA-uUskxIAEks5qUn2wgaJpZM4JJfoa>
> .
>