@jmxpearson 2016-07-14
This looks great!

A couple of questions:

1. Does this approach rule out the matrix optimizations we have on master? 

    From @mlubin, I understand that in the case of JuMP, there are often huge numbers of variables with thousands of constraints, all distinct. But for many of the ML applications I want, the graph is not terribly large; many of the nodes, though are substantial tensors. The approach of using an elementwise tape on these will be pretty inefficient, I think, since it prevents Julia from doing things like using optimized linear algebra for backprop.

2. I've thought once again about trying to spec out the codegen approach, this time allowing variables to be arrays in addition to scalars. This should cut down generated code a lot and avoid the bloat I feared initially. Have you thought much about this? Is there something simple I'm missing?