@jrevels 2016-07-14
> Does this approach rule out the matrix optimizations we have on master?

I've been thinking hard about this one as well. If we were to continue with this approach, we'd need different wrapper types for the different functions we'd like to optimize. Specifically, I'm thinking about `TraceArray <: AbstractArray` for array-to-array functions (like `map`) and `TraceReduction <: AbstractTraceReal <: Real` for array-to-scalar operations (like `sum`). Then, we'd implement `backprop_step!` on these types. 

I do wonder if there's a way to combine this PR's approach and the approach on `master`, i.e. leveraging `immutable` + `Ref` + forward derivative accumulation with `Dual`s to achieve a non-global tape-based storage that is easier to traverse and optimize array methods for. This thought is what's kept me from hitting merge here.

> I've thought once again about trying to spec out the codegen approach, this time allowing variables to be arrays in addition to scalars. This should cut down generated code a lot and avoid the bloat I feared initially. Have you thought much about this? Is there something simple I'm missing?

I've only thought about the compiled approach a little bit versus the interpreted approach, so I have no idea what the hurdles would be there. Sounds like it's worth trying, though.

I was thinking another big win for the compiled approach could come from partitioning the DAG into smaller subsections that could be compiled as smaller Julia functions, which would probably lead to better compile-time and run-time performance. We could accumulate info about function barriers during the trace that might make this feasible. The real trick with the compiled approach would be not re-compiling for data-independent (i.e. tree-representable) subsections of the DAG. With that in mind, it might be that the compiled approach could only work well for the case described in #11.