@jrevels 2016-08-09
superceded by #13